feat: 封装组合输入控件，减少重复代码
1.封装选择文件输入控件，默认打开文件夹，可传入自定义逻辑
2.新控件替换现有实现，删除重复代码

+++ app/codec/ffmpegdecoder.cc
@@ -331,7 +331,7 @@ bool FFmpegDecoder::InitInputFmtParams(std::string& url, AVInputFormat** fmt)
         break;
     }
     default:
-        SPDLOG_ERROR("Failed to open this media type {0}.", media_->type);
+        SPDLOG_ERROR("Failed to open this media type {0}.", (int)media_->type);
         return false;
     }
